test(stripe): split webhook scenarios into webhook.test.ts
Pulls 24 webhook handler tests (5 pre-describe events tests, 9 skip
and dedup tests, 5 schedule sync tests, plus Webhook Error Handling
on Stripe v19, cancel_at_period_end, and immediate cancellation
describes) out of stripe.test.ts. Wraps them in a "stripe webhook"
describe. stripe.test.ts loses ~2,720 lines.
